Move reranking plugin under core/ml
