+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/reranking/TextSimilarityRankDoc.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.rank;
+package org.elasticsearch.xpack.core.ml.reranking;
 
 import org.elasticsearch.common.io.stream.StreamInput;
 import org.elasticsearch.common.io.stream.StreamOutput;
