+++ x-pack/plugin/core/src/main/java/org/elasticsearch/xpack/core/ml/reranking/TextSimilarityRankShardResult.java
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-package org.elasticsearch.xpack.inference.rank;
+package org.elasticsearch.xpack.core.ml.reranking;
 
 import org.elasticsearch.TransportVersion;
 import org.elasticsearch.TransportVersions;
