--- x-pack/plugin/inference/src/main/resources/META-INF/services/org.elasticsearch.features.FeatureSpecification
@@ -1,8 +1,0 @@
-#
-# Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-# or more contributor license agreements. Licensed under the Elastic License
-# 2.0; you may not use this file except in compliance with the Elastic License
-# 2.0.
-#
-
-org.elasticsearch.xpack.inference.InferenceFeatures
